Linked user results from search bar to show the modules taken by those respective users. Profile page also updated to include the three fields for a user
