+++ src/components/auth/AuthTextInput.js
@@ -7,7 +7,7 @@ const THEME = 'lightgrey';
  const AuthTextInput = props => {
 
     
-    const [secureTextEntry, setSecureTextEntry] = React.useState(true);
+    const [secureTextEntry, setSecureTextEntry] = React.useState(false);
 
     const { placeholder, keyboardType, value, textHandler } = props;
