+++ src/components/index.js
@@ -1,5 +1,6 @@
 export { default as AuthTextInput } from './auth/AuthTextInput';
 export { default as AuthPressable } from './auth/AuthPressable';
+export { default as LogoutButton } from './auth/LogoutButton';
 export { default as Module } from './module/Module';
 export { default as UserResult }  from './search/UserResult';
 export { default as DATA } from './search/DATA';
