+++ src/screens/ProfileScreen.js
@@ -1,4 +1,3 @@
-import { View} from 'react-native'
 import React from 'react'
 import { 
     Divider, 
@@ -11,31 +10,58 @@ import {
     Text,
 } from '@ui-kitten/components';
 import {
-    HomeScreen,
-    SearchScreen,
+    ModuleScreen,
 } from '../screens';
-import { KeyboardAvoidingView, SafeAreaView, StyleSheet,} from "react-native";
+import { KeyboardAvoidingView, SafeAreaView,} from "react-native";
+import { LogoutButton, UserResult } from '../components';
+import { auth, db } from '../firebase';
+import {
+    doc,
+    getDoc,
+    setDoc,
+} from 'firebase/firestore';
+import { useEffect } from 'react';
 
 const ProfileScreen= () => {
-  return (
-    <SafeAreaView style={{flex:1}}>
-    <KeyboardAvoidingView style={{flex:1}}>
-        <TopNavigation 
-            title='Profile'
-            alignment='start'
-            accessoryLeft={<Icon name='frienus' pack='customAssets' style={{marginLeft:5, height:60, width:60}} />}
-            style={{height:'8%'}}
-        />
-        <Divider/>
-        <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text category='h1'>Profile</Text>
-        <Button onPress={() => HomeScreen}>
-            BUTTON
-        </Button>
-        </Layout>
-    </KeyboardAvoidingView>
-    </SafeAreaView>
-  )
+    const [displayNameField, setDisplayNameField] = React.useState("");
+    const [emailField, setEmailField] = React.useState("");
+    const [idField, setIdField] = React.useState("");
+
+    useEffect(() => {
+        const currUser = auth.currentUser;
+        const userDoc = doc(db, 'Users/' + currUser.uid);
+        const docSnap = getDoc(userDoc).then(result => {
+            console.log("This is the currentUser id: " + currUser.uid);
+            setDisplayNameField(result.get('displayName'));
+            setEmailField(result.get('email'));
+            setIdField(result.get('id'));
+        })
+    })
+
+    
+    return (
+        <SafeAreaView style={{flex:1}}>
+        <KeyboardAvoidingView style={{flex:1}}>
+            <TopNavigation 
+                title='Profile'
+                alignment='start'
+                accessoryLeft={<Icon name='frienus' pack='customAssets' style={{marginLeft:5, height:60, width:60}} />}
+                accessoryRight={LogoutButton}
+                style={{height:'8%'}}
+            />
+            <Divider/>
+            <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+            <Text category='h1'>Profile</Text>
+            <Button onPress={() => ModuleScreen}>
+                BUTTON
+            </Button>
+            <Text>This is your current uid: {idField}</Text>
+            <Text>Your email is now: {emailField} </Text>
+            <Text>Your display name is: {displayNameField} </Text>
+            </Layout>
+        </KeyboardAvoidingView>
+        </SafeAreaView>
+    )
 }
 
 export default ProfileScreen;
