+++ src/screens/index.js
@@ -1,4 +1,4 @@
 export { default as AuthScreen } from './AuthScreen';
-export { default as HomeScreen } from './HomeScreen';
+export { default as ModuleScreen } from './ModuleScreen';
 export { default as SearchScreen } from './SearchScreen';
 export { default as ProfileScreen } from './ProfileScreen';
